Replace CvExcept with MonadError
This commit removes CvExcept and CvExceptT, and rewrites all OpenCV functions to
use a polymorphic MonadError constraint instead. This has the benefit of easing
the use of the library for users, as they no longer have to worry about lifting
operations (lifting CvExcept into CvExceptT if they are mixing pure and impure
operations), and also removes the cognitive burden of figuring out what
'CvExcept' is and how to work with it.

Also, by being polymorphic in the error representation, we potentially make it
easier to write "one off" OpenCV function calls. Either CvException is an
instance of MonadError CvException, so people can simply pattern match on the
result of an OpenCV call to witness Left/Right constructors, rather than having
to jump through the hoop of runExcept (which incur a transformers dependency -
minor, but annoying overhead).