prevent allocaBytes with 0 bytes
The function withArrayPtr is unsafe when applied to an empty
vector. We allocate memory with allocaBytes for 0 bytes and then use
the resulting pointer which can cause segmentation faults.

Many parts of the code are susceptible to this bug. It is fixed by
changing the behaviour of the withArrayPtr function to only call the
inner computation when the supplied vector is non empty.