Mat: Refactor: Switch from `mkFinalizer` to inline-c's `funPtr`.
`funPtr` was added to `inline-c` specifically for this purpose and I believe
that it is slightly better than the current TH solution, simply because it
is more obvious what's going on.

In any case, this commit switches only one of multiple occurrences of
`mkFinalizer` over to the new `funPtr`.

This patch was originally written for version 0.0.2.1, which used
`Foreign.Concurrent.newForeignPtr`, which was a problem.
The original commit message of this patch was (retained for information):

		Mat: Use new inline-c funPtr to make ForeignPtr be freed more reliably.

		This removes a use of `objFromPtr` which removes a use of
		`Foreign.Concurrent`, which makes freeing not dependent on a thread.

		See https://github.com/fpco/inline-c/issues/77 and
		https://github.com/fpco/inline-c/commit/2779d6ef81cb1b022ab8da6f4d8cf8a5a3194a8c
		 for details on why `Foreign.Concurrent.newForeignPtr` is worse than normal
		`newForeignPtr`.