Use size_t instead of int32_t where appropriate
As reported in #133 there where a number of places taking a values
like vector.size() which is of type size_t and storing it in a int32_t
variable (via an implicit conversion).

These places are now changed to work with size_t variables in C and
CSize in Haskell. Still we convert most of these to Int (because we
often need it as an argument to peekArray which takes an Int). So on
platforms in which Int is smaller then CSize there is still the
potential for overflow errors.