Fix #13: runtime crash of opencv-examples
The opencv3/4 package in nixpkgs by default doesn't enable GTK
support. This causes the executables in the opencv-examples and
opencv-extra-examples Haskell packages to fail because they are using
the highgui OpenCV module:

terminate called after throwing an instance of 'cv::Exception'
  what():  OpenCV(3.4.5)
  /build/source/modules/highgui/src/window.cpp:617:
    error: (-2:Unspecified error) The function is not implemented.
    Rebuild the library with Windows, GTK+ 2.x or Carbon support.
    If you are on Ubuntu or Debian, install libgtk2.0-dev
    and pkg-config, then re-run cmake or configure script
    in function 'cvNamedWindow'

To fix this we override the opencv3/4 package to enable GTK
support. However we only override it for the opencv-examples and
opencv-extra-examples packages. This means you don't have to rebuild
opencv3/4 when you're just working on the opencv or opencv-extra
Haskell packages and can download the package from the cache.

Note that on OS X the opencv3/4 package by default has a working
highgui module so on that platform we don't need to enable GTK support.